WIP: Implement feature Optimistic Concurrency Support via Concurrency Tokens
