+++ benchmarks/DbConnectionPlus.Benchmarks/Benchmarks.DeleteEntity.cs
@@ -1,0 +1,81 @@
+// ReSharper disable InvokeAsExtensionMethod
+// ReSharper disable InconsistentNaming
+
+#pragma warning disable RCS1196
+
+namespace RentADeveloper.DbConnectionPlus.Benchmarks;
+
+public partial class Benchmarks
+{
+    [IterationCleanup(
+        Targets =
+        [
+            nameof(DeleteEntity_DbCommand),
+            nameof(DeleteEntity_Dapper),
+            nameof(DeleteEntity_DbConnectionPlus)
+        ]
+    )]
+    public void DeleteEntity__Cleanup() =>
+        this.SetupDatabase(DeleteEntity_OperationsPerInvoke);
+
+    [IterationSetup(
+        Targets =
+        [
+            nameof(DeleteEntity_DbCommand),
+            nameof(DeleteEntity_Dapper),
+            nameof(DeleteEntity_DbConnectionPlus)
+        ]
+    )]
+    public void DeleteEntity__Setup() =>
+        this.SetupDatabase(DeleteEntity_OperationsPerInvoke);
+
+    [Benchmark(Baseline = false, OperationsPerInvoke = DeleteEntity_OperationsPerInvoke)]
+    [BenchmarkCategory(DeleteEntity_Category)]
+    public void DeleteEntity_Dapper()
+    {
+        for (var i = 0; i < DeleteEntity_OperationsPerInvoke; i++)
+        {
+            var entityToDelete = this.entitiesInDb[0];
+
+            SqlMapperExtensions.Delete(this.connection, entityToDelete);
+
+            this.entitiesInDb.Remove(entityToDelete);
+        }
+    }
+
+    [Benchmark(Baseline = true, OperationsPerInvoke = DeleteEntity_OperationsPerInvoke)]
+    [BenchmarkCategory(DeleteEntity_Category)]
+    public void DeleteEntity_DbCommand()
+    {
+        for (var i = 0; i < DeleteEntity_OperationsPerInvoke; i++)
+        {
+            var entityToDelete = this.entitiesInDb[0];
+
+            using var command = this.connection.CreateCommand();
+
+            command.CommandText = "DELETE FROM Entity WHERE Id = @Id";
+            command.Parameters.Add(new("@Id", entityToDelete.Id));
+
+            command.ExecuteNonQuery();
+
+            this.entitiesInDb.Remove(entityToDelete);
+        }
+    }
+
+    [Benchmark(Baseline = false, OperationsPerInvoke = DeleteEntity_OperationsPerInvoke)]
+    [BenchmarkCategory(DeleteEntity_Category)]
+    public void DeleteEntity_DbConnectionPlus()
+    {
+        for (var i = 0; i < DeleteEntity_OperationsPerInvoke; i++)
+        {
+            var entityToDelete = this.entitiesInDb[0];
+
+            this.connection.DeleteEntity(entityToDelete);
+
+            this.entitiesInDb.Remove(entityToDelete);
+        }
+    }
+
+    private const String DeleteEntity_Category = "DeleteEntity";
+    private const Int32 DeleteEntity_OperationsPerInvoke = 8000;
+}
