+++ benchmarks/DbConnectionPlus.Benchmarks/BenchmarksConfig.cs
@@ -4,7 +4,6 @@ using BenchmarkDotNet.Exporters;
 using BenchmarkDotNet.Jobs;
 using BenchmarkDotNet.Reports;
 using Perfolizer.Horology;
-using Perfolizer.Mathematics.OutlierDetection;
 
 namespace RentADeveloper.DbConnectionPlus.Benchmarks;
 
@@ -16,25 +15,14 @@ public class BenchmarksConfig : ManualConfig
         this.Orderer = new BenchmarksOrderer();
         this.SummaryStyle = SummaryStyle.Default.WithRatioStyle(RatioStyle.Trend);
 
-        this.AddColumn(StatisticColumn.Median);
-        this.AddColumn(StatisticColumn.P90);
-        this.AddColumn(StatisticColumn.P95);
-
-        this.AddExporter(PlainExporter.Default);
         this.AddExporter(MarkdownExporter.Default);
 
         this.AddJob(
             Job.Default
-                .WithWarmupCount(10)
-                .WithMinIterationTime(TimeInterval.FromMilliseconds(100))
-                .WithMaxIterationCount(20)
-                .WithInvocationCount(1)
-                .WithUnrollFactor(1)
+                .WithMinIterationTime(TimeInterval.FromMilliseconds(150))
 
                 // Since DbConnectionPlus will mostly be used in server applications, we test with server GC.
                 .WithGcServer(true)
-
-                .WithOutlierMode(OutlierMode.DontRemove)
         );
     }
 }
