+++ benchmarks/DbConnectionPlus.Benchmarks/DapperTypeHandlers/GuidTypeHandler.cs
@@ -1,14 +1,12 @@
-﻿using Dapper;
-
-namespace RentADeveloper.DbConnectionPlus.Benchmarks.DapperTypeHandlers;
+﻿namespace RentADeveloper.DbConnectionPlus.Benchmarks.DapperTypeHandlers;
 
 public class GuidTypeHandler : SqlMapper.StringTypeHandler<Guid>
 {
     /// <inheritdoc />
-    protected override Guid Parse(String xml) =>
-        Guid.Parse(xml);
-
-    /// <inheritdoc />
     protected override String Format(Guid xml) =>
         xml.ToString();
+
+    /// <inheritdoc />
+    protected override Guid Parse(String xml) =>
+        Guid.Parse(xml);
 }
