+++ benchmarks/DbConnectionPlus.Benchmarks/GlobalUsings.cs
@@ -1,0 +1,10 @@
+global using System.Data;
+global using System.Globalization;
+global using BenchmarkDotNet.Attributes;
+global using Dapper;
+global using Dapper.Contrib.Extensions;
+global using Microsoft.Data.Sqlite;
+global using RentADeveloper.DbConnectionPlus.Benchmarks.DapperTypeHandlers;
+global using RentADeveloper.DbConnectionPlus.Benchmarks.TestData;
+global using RentADeveloper.DbConnectionPlus.UnitTests.TestData;
+global using static RentADeveloper.DbConnectionPlus.DbConnectionExtensions;
